requirements.txt
langchain
ipykernel
python_dotenv
langchain-community
pypdf
langchain-text-splitters
langchain-google-genai
langchain-ollama
langchain-huggingface
sentence_transformers
faiss-cpu
streamlit
langchain_groq
pydantic==2.10.6
langgraph
langchain-tavily
langgraph
langchain-huggingface
arxiv
tavily-python
langchain-tavily